requirements.txt
streamlit==1.36.0
Pillow==10.0.1
pytesseract==0.3.10
pyperclip==1.9.0